+++ src/com/amazon/speech/speechlet/PlaybackStoppedRequest.java
@@ -10,7 +10,7 @@ import com.fasterxml.jackson.annotation.JsonTypeName;
 @JsonTypeName("AudioPlayer.PlaybackStopped")
 public final class PlaybackStoppedRequest extends SpeechletRequest {
     private final String token;
-    private final int offsetInMilliseconds;
+    private final long offsetInMilliseconds;
     private final String locale;
 
     public static Builder builder() {
@@ -28,7 +28,7 @@ public final class PlaybackStoppedRequest extends SpeechletRequest {
             @JsonProperty("requestId") final String requestId,
             @JsonProperty("timestamp") final Date timestamp,
             @JsonProperty("token") final String token,
-            @JsonProperty("offsetInMilliseconds") final int offsetInMilliseconds,
+            @JsonProperty("offsetInMilliseconds") final long offsetInMilliseconds,
             @JsonProperty("locale") final String locale
     ) {
         super(requestId, timestamp);
@@ -41,7 +41,7 @@ public final class PlaybackStoppedRequest extends SpeechletRequest {
         return this.token;
     }
 
-    public int getOffsetInMilliseconds() {
+    public long getOffsetInMilliseconds() {
         return this.offsetInMilliseconds;
     }
 
@@ -53,7 +53,7 @@ public final class PlaybackStoppedRequest extends SpeechletRequest {
         private String requestId;
         private Date timestamp = new Date();
         private String token;
-        private int offsetInMilliseconds;
+        private long offsetInMilliseconds;
         private String locale;
 
         private Builder() {
@@ -74,7 +74,7 @@ public final class PlaybackStoppedRequest extends SpeechletRequest {
             return this;
         }
 
-        public Builder withOffsetInMilliseconds(final int offsetInMilliseconds) {
+        public Builder withOffsetInMilliseconds(final long offsetInMilliseconds) {
             this.offsetInMilliseconds = offsetInMilliseconds;
             return this;
         }
